microservices/car_tracker/requirements.txt
flask
ultralytics
kafka-python
opencv-python
numpy
pandas
scipy